+++ factory-cli/src/services/source/typescript/app-configuration/app-configuration.interface.ts
@@ -1,0 +1,8 @@
+import { SolutionProject } from "../../../../models/solution";
+
+export interface IAppConfiguration {
+  configType: string;
+  getConfigName(): string;
+  addAppConfigurationJson(solutionFolderPath: string, project: SolutionProject): Promise<void>;
+  addAppConfigurationModel(solutionFolderPath: string, project: SolutionProject): Promise<void>;
+}
