+++ factory-cli/src/transformations/node/node-compose-datacontext.transform.ts
@@ -1,11 +1,11 @@
 import { ProjectTransformation, Solution } from "../../models/solution";
 import { FileService, IFileService } from "../../services/file.service";
-import { ITypescriptSourceService, TypescriptSourceService } from "../../services/source/typescript-source.service";
+import { ITypescriptSourceService, TypescriptSourceService } from "../../services/source/typescript/typescript-source.service";
 import { ITransformation } from "../transformation";
 
-export class NodeComposeSqliteDataContextTransformation implements ITransformation {
+export class NodeComposeDataContextTransformation implements ITransformation {
 
-  get name(): string { return "compose:datacontext-sqlite"; }
+  get name(): string { return "compose:datacontext"; }
   get environment(): string { return "node"; }
   get language(): string { return "typescript"; }
   fileService: IFileService = new FileService();
@@ -16,12 +16,14 @@ export class NodeComposeSqliteDataContextTransformation implements ITransformati
     if (!project) return Promise.reject(new Error(`Invalid target project in transformation: '${transformation.targetProject}'.`));
     let databaseProject = solution.projects.find(p => p.name == transformation.sourceProject);
     if (!databaseProject) return Promise.reject(new Error(`Invalid source project in transformation: '${transformation.sourceProject}'.`));
+    let databaseType = databaseProject.specs?.databaseType;
+    if (!databaseType) return Promise.reject(new Error("Database type not specified in database project specs."));
     const contextName = databaseProject.specs?.contextName ?? "SampleDatabaseContext";
     return this.sourceService.createAppConfigFromSampleConfig(solutionFolderPath, project)
-      .then(_ => this.sourceService.addSqliteAppConfigurationJson(solutionFolderPath, project))
-      .then(_ => this.sourceService.addSqliteAppConfigurationModel(solutionFolderPath, project))
+      .then(_ => this.sourceService.addAppConfigurationJson(solutionFolderPath, project, databaseType))
+      .then(_ => this.sourceService.addAppConfigurationModel(solutionFolderPath, project, databaseType))
       .then(_ => this.sourceService.addDataContextCompositionType(solutionFolderPath, project, contextName))
-      .then(_ => this.sourceService.addDataContextComposition(solutionFolderPath, project, databaseProject.name, contextName, "sqliteConfig"));
+      .then(_ => this.sourceService.addDataContextComposition(solutionFolderPath, project, databaseProject.name, contextName, databaseType));
   }
 
 }
